Move serialization functions to a different package
